version 2.0.5
While importing an image, you can now select "Nearest Multiple of 4", which is the suggested bare-minimum for mipmaps to be generated properly.

If you have an image opened, you can now change the version by selecting the new one in the dropdown box in the Info tab, then Save. (Keep in mind this will change the VTF size. So if you are using it for a spray decal, it could possibly go outside of the 512kb limitation when increasing the version.)

The program's settings file is now contained in the same location that the program is in, instead of your Roaming folder.

The default version has also been changed to 7.4 as mostly commonly that will be the version used by source engine modders.

The context menu's options for VTF files has been slightly changed for more clarity.

The File menu has been slightly changed for more clarity and better positioning. Creating either VTF or VMT is now under the first option "Create New..."

A pull request by goodusername123 in the main branch has also been made to this repo, which excludes the sRGB flag from VTFs. This is hopefully a temporary measure.

Reverted the syntax highlighter as it's terrible with performance and breaks some functionality. All related code is commented out if the issues can be solved later.

Added a way to change the framerate for animated playback (in milliseconds).

Fixed the import's resize filtering to use what you chose instead of being forced to use only one type of filter.